Render iOS auth SDK errors specifically
