+++ Packages/iOS/CmuxMobileShellUI/Package.swift
@@ -61,6 +61,7 @@ let package = Package(
                 "CmuxMobileShellUI",
                 "CmuxMobileShell",
                 "CmuxMobileWorkspace",
+                .product(name: "StackAuth", package: "stack-auth-swift-sdk-prerelease"),
             ],
             swiftSettings: [
                 .define("CMUX_DEV_AUTH", .when(configuration: .debug)),
